requirements.txt
requests
openai-whisper
fastapi
uvicorn[standard]
pydantic
python-dotenv
loguru